Bugfix for JaggZilla bug #515
When we fly to a sector with enemies and assign at least one of our passengers in the heli to a squad the game will remove everyone from the heli and start combat. This prevents the exploits that existed before where we could assign our men to anything in hostile sectors.

git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk@6486 3b4a5df2-a311-0410-b5c6-a8a6f20db521